extract and test isXsdInteger
