+++ lib/SerializerStream.js
@@ -1,6 +1,7 @@
 const ObjectEncoder = require('./ObjectEncoder')
 const Readable = require('readable-stream')
 const StringEncoder = require('./StringEncoder')
+const isXsdInteger = require('./isXsdInteger')
 
 class SerializerStream extends Readable {
   constructor (input, {encoding = 'object'} = {}) {
@@ -63,27 +64,13 @@ class SerializerStream extends Readable {
     if (object.language) {
       return {'@language': object.language, '@value': object.value}
     } else if (isXsdInteger(object)) {
-      return parseInt(object.value)
+      return parseInt(object.value, 10)
     } else if (object.datatype && object.datatype.value !== 'http://www.w3.org/2001/XMLSchema#string') {
       return {'@type': object.datatype.value, '@value': object.value}
     } else {
       return object.value
     }
   }
-}
-
-function isXsdInteger (object) {
-  return object.datatype &&
-    isXsdIntegerType(object.datatype) &&
-    isIntegerValue(object.value)
-}
-
-function isXsdIntegerType (datatype) {
-  return datatype.value === 'http://www.w3.org/2001/XMLSchema#integer'
-}
-
-function isIntegerValue (value) {
-  return !isNaN(value) && parseInt(value).toFixed(0) === value
 }
 
 module.exports = SerializerStream
